Fix enso italic mode for Comments on cterm=
Not all terminal emulators support italic, so be safe and turn this off.
Worked okay with iTerm, and showed italics, but with Terminal.app it
didn't show italics. Even worse, with Terminal and tmux, even with
screen-256color set as the $TERM value in .tmux.conf, italics would not
work but also messed up the colors for Comments highlight group.

TLDR: italics are finnicky in cterm, and prone to break/miscolor Comment
highlight group, so turn them off.