Attempt at includexpr for vim files.
Unfortunately gf behaviour is to jump to directory is if matches the
same pattern as the filename, e.g., plugin/pack.vim/pack.vim would open
netrw in pack.vim/ directory, not the file. Only way around this is
remap gf to custom function to only check for files matching the given
pattern.