feat: Use syntax for omnifunc after FileType set.
Using autocmd to detect FileType being set, then set omnifunc to
complete using syntaxcomplete#Complete (shipped with Vim) if an omnifunc
was not already defined by the ftplugin runtime files.

This helps with covering gaps that YCM does not cover, for instance
there without this you get nothing on <C-Space> with YCM because it
doesn't have a vim completer.