feat(italics): Add terminfo file to support italics.
On Terminal.app we need to set our profile to run:
export TERM=xterm-256color-italic
on startup, to tell it we *can* use italics, because the macOS
terminfo is so old it doesn't have the supported lines in it.

To 'install' this new terminfo, we just do:
tic -x ~/.vim/xterm-256color-italic.terminfo

And it will be put into a ~/.terminfo/.. where terminfo will look
for it. The file itself just defines the entries needed to support
italics, then sources the basic xterm-256color terminfo.